requirements.txt
flask[async]
flask-session 
google-generativeai 
markdown
pandas
PyPDF2
flask_cors
python-dotenv
PyMuPDF
sentence-transformers
faiss-cpu
nest_asyncio



